refactor: Rename types to cell-types
