+++ packages/neuron-wallet/src/services/sync-blocks.ts
@@ -2,7 +2,7 @@ import { getConnection } from 'typeorm'
 import { remote } from 'electron'
 import { Subject, BehaviorSubject } from 'rxjs'
 import Core from '@nervosnetwork/ckb-sdk-core'
-import { Script, OutPoint, Cell, Input, Transaction, Block, BlockHeader } from '../types/types'
+import { Script, OutPoint, Cell, Input, Transaction, Block, BlockHeader } from '../types/cell-types'
 import TransactionsService from './transactions'
 import OutputEntity from '../database/chain/entities/output'
 import SyncInfoEntity from '../database/chain/entities/sync-info'
