+++ packages/neuron-wallet/src/services/wallets.ts
@@ -10,7 +10,7 @@ import FileService from './file'
 import LockUtils from '../utils/lock-utils'
 import windowManager from '../models/window-manager'
 import { Channel, ResponseCode } from '../utils/const'
-import { Witness, TransactionWithoutHash, Input } from '../types/types'
+import { Witness, TransactionWithoutHash, Input } from '../types/cell-types'
 import ConvertTo from '../types/convert-to'
 import Blake2b from '../utils/blake2b'
 import { CurrentWalletNotSet, WalletNotFound, IsRequired, UsedName } from '../exceptions'
