+++ packages/neuron-wallet/tests/services/wallets.test.ts
@@ -1,7 +1,7 @@
 import '../../src/controllers/wallets'
 import '../../src/controllers/app'
 import WalletService, { WalletProperties } from '../../src/services/wallets'
-import { Witness } from '../../src/types/types'
+import { Witness } from '../../src/types/cell-types'
 import Keystore from '../../src/models/keys/keystore'
 import Keychain from '../../src/models/keys/keychain'
 import { mnemonicToSeedSync } from '../../src/models/keys/mnemonic'
